Creat data model and Example jdp to call backend
