+++ src/main/java/edu/uwaterloo/ece658/entity/LocationTag.java
@@ -5,11 +5,7 @@
  */
 package edu.uwaterloo.ece658.entity;
 
-import java.io.Serializable;
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
 import javax.persistence.Table;
 
 /**
@@ -17,12 +13,12 @@ import javax.persistence.Table;
  * @author mier
  */
 @Entity
-@Table(name="T_LOCATION_TAG")
+@Table(name = "T_LOCATION_TAG")
 public class LocationTag extends Tag {
 
     @Override
     public String toString() {
         return "edu.uwaterloo.ece658.entity.LocationTag[ id=" + id + " ]";
     }
-    
+
 }
