+++ src/main/java/edu/uwaterloo/ece658/entity/TimeTag.java
@@ -5,11 +5,7 @@
  */
 package edu.uwaterloo.ece658.entity;
 
-import java.io.Serializable;
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
 import javax.persistence.Table;
 
 /**
@@ -17,12 +13,12 @@ import javax.persistence.Table;
  * @author mier
  */
 @Entity
-@Table(name="T_TIME_TAG")
+@Table(name = "T_TIME_TAG")
 public class TimeTag extends Tag {
 
     @Override
     public String toString() {
         return "edu.uwaterloo.ece658.entity.TimeTag[ id=" + id + " ]";
     }
-    
+
 }
