--- src/main/java/edu/uwaterloo/ece658/servlet/README.md
@@ -1,1 +1,0 @@
-Add Servlet classes that handle web requests here.
