+++ src/main/resources/META-INF/persistence.xml
@@ -1,12 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <persistence version="2.0" xmlns="http://java.sun.com/xml/ns/persistence" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://java.sun.com/xml/ns/persistence   http://java.sun.com/xml/ns/persistence/persistence_2_0.xsd">
-  <persistence-unit name="PhototrestPU" transaction-type="JTA">
-    <provider>com.objectdb.jpa.Provider</provider>
-    <properties>
-<!--      <property name="objectdb" value="/home/mier/Downloads/objectdb-2.7.2_0"/>-->
-      <property name="javax.persistence.jdbc.url" value="/db/ece658.odb"/>
-      <property name="javax.persistence.jdbc.user" value="admin"/>
-      <property name="javax.persistence.jdbc.password" value="admin"/>
-    </properties>
-  </persistence-unit>
+    <persistence-unit name="PhototrestPU" transaction-type="JTA">
+        <jta-data-source>java:app/derby</jta-data-source>
+        <properties>
+            <property name="eclipselink.ddl-generation" value="drop-and-create-tables"/>
+            <property name="eclipselink.logging.level" value="INFO"/>
+            <property name="eclipselink.logging.level.sql" value="INFO"/>
+        </properties>
+    </persistence-unit>
 </persistence>
